Add non-plugin build into the action
Several variables have been added to the build configuration. The dockerfile variable is needed to dynamically determine which dockerfile will be used for the build, and a variable was added to the build tag to push images without plugins to the repository. Added a matrix for building versions without plugins. Added a name for matrices for a more visual result.